+++ src/main/java/testapp/models/City.java
@@ -8,19 +8,19 @@ public class City implements ILuwakModel {
 	public static final String NAME = "name";
 	public static final String COUNTRY = "country";
 
-    private int id;
+    private long id;
     private String name;
     private Country country;
 
     public City() {}
 
-    public City(int id, String name, Country country) {
+    public City(long id, String name, Country country) {
     	this.id = id;
     	this.name = name;
     	this.country = country;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
