+++ src/main/java/testapp/models/Country.kt
@@ -2,7 +2,7 @@ package testapp.models
 
 import com.menighin.luwak.core.interfaces.ILuwakModel
 
-class Country(var id: Int, var code: String?) : ILuwakModel {
+class Country(var id: Long, var code: String?) : ILuwakModel {
 
     companion object {
         const val ID: String = "id"
