+++ src/main/java/testapp/models/CountryData.kt
@@ -2,7 +2,7 @@ package testapp.models
 
 import com.menighin.luwak.core.interfaces.ILuwakModel
 
-data class CountryData(val id: Int, val country: Country, val key: String, val value: Int) : ILuwakModel {
+data class CountryData(val id: Long, val country: Country, val key: String, val value: Int) : ILuwakModel {
 
 	companion object {
 		const val ID: String = "id"
